requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
Django==4.2.11
django-environ==0.11.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lxml==5.3.1
openai==1.10.0
packaging==24.2
pdf2image==1.16.3
Pillow==10.0.0
pydantic==2.10.6
pydantic_core==2.27.2
PyPDF2==3.0.1
pytesseract==0.3.10
python-docx==0.8.11
requests==2.31.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0